requirements.txt
asgiref==3.8.1
click==8.1.7
colorama==0.4.6
cssbeautifier==1.15.1
Django==5.1.1
django-admin-autocomplete-filter==0.7.1
django-debug-toolbar==4.4.6
django-js-asset==2.2.0
django-modeltranslation==0.19.10
django-mptt==0.16.0
djlint==1.35.2
EditorConfig==0.12.4
html-tag-names==0.1.2
html-void-elements==0.1.0
jsbeautifier==1.15.1
json5==0.9.25
pathspec==0.12.1
pillow==10.4.0
PyYAML==6.0.2
regex==2024.9.11
six==1.16.0
sqlparse==0.5.1
tomli==2.0.2
tqdm==4.66.5
typing_extensions==4.12.2
